+++ lib/l10n/translations/pl.dart
@@ -1788,6 +1788,7 @@ const _dutchSourcePl = {
 };
 
 const _dutchSourceAddPl = <String, String>{
+  'De export is mislukt.': 'Eksport nie powiódł się.',
   'Teams-gastclient (ontwerp)': 'Klient Teams dla gości (projekt)',
   'Traffic Light Protocol: hoe breed mag dit materiaal gedeeld worden?':
       'Traffic Light Protocol: jak szeroko można udostępniać ten materiał?',